Moved stringtables stuff to the appropriate files
Fixed StringTable.changed_since_tick() calling the wrong C++ function
